Build clean ups and warning squashing
Cleaning up the builds so both Debug and Release, and x86 and x64 processes work correctly.

Getting rid of typing issues with standard library size_t differences.

+++ StrangeGameEngine/render.cpp
@@ -1081,7 +1081,7 @@ namespace SGE
 		int RenderBitmapFile::LoadFile(char* targetFilename)
 		{
 			FILE* bitmapFile;
-			unsigned int readCount = 0;
+			size_t readCount = 0;
 			unsigned char pixelRed;
 			unsigned char pixelGreen;
 			unsigned char pixelBlue;
